remove / CREDIT REPORT DISPUTE description text
- What I started filling out the credit report tool, I still see the
text at the top that says: A little known fact of the debt collections
industry is that the vast majority of collectors can’t prove they own
our debts. If they can’t prove it, then why should we pay? Demanding
proof of ownership is the first step to getting debt collectors off our
backs.

This text is not necessary for this tool. It can be deleted.