fix(DebtAmounts): update component constrain to accept string type on amount
Before this fix the values were handled with strings, and while makes sense to push the amount as a
number it adds more complexity to the final solution. After reach the point to decide what to do as
commented https://github.com/debtcollective/dispute-tools/issues/169#issuecomment-484873701 the
decision is to keep the types as string. Nevertheless, the vue component will check if it is a valid
number before sending the value and we also have a mask input on top of that to enforce formatting

fix dispute-tools#169